+++ client/SparseMatrix.h
@@ -11,8 +11,8 @@ private:
     int numRows, numCols, numValues;
 
     std::vector<int> rowIds, colIds;
-    std::vector<float> values;
-    std::vector<float> b;
+    std::vector<double> values;
+    std::vector<double> b;
 
     std::ifstream matrix;
 public:
@@ -25,11 +25,11 @@ public:
     int getNumberOfValues() const { return numValues; }
     int* getRowIds() { return rowIds.data(); }
     int* getColIds() { return colIds.data(); }
-    float* getValues() { return values.data(); }
-    float* getVectorB() { return b.data(); }
+    double* getValues() { return values.data(); }
+    double* getVectorB() { return b.data(); }
 
-    void fillVectorBWithRandomValues(float minValue, float maxValue);
-    void fillVectorBFullyWithConcreteValue(float value);
+    void fillVectorBWithRandomValues(double minValue, double maxValue);
+    void fillVectorBFullyWithConcreteValue(double value);
 };
 
 #endif //COURSE_WORK_SOLVE_SYSTEM_OF_SPARSE_LINEAR_EQUATIONS_SPARSEMATRIX_H
